+++ src/main/java/memethespire/relicmemes/CombatRewardScreenSetupItemRewardPatch.java
@@ -13,7 +13,8 @@ public class CombatRewardScreenSetupItemRewardPatch {
     public static void modifyRelics(CombatRewardScreen instance) {
         instance.rewards.forEach((reward) -> {
             if (reward.type == RewardItem.RewardType.RELIC) {
-                MemeCollection.applyFirstApplicableRelicModificationFromAllCollections(reward.relic, AbstractDungeon.player);
+                MemeCollection.applyFirstApplicableRelicModificationFromAllCollections(
+                        reward.relic, AbstractDungeon.player, false);
                 // Update the name on the reward item as well.
                 reward.text = reward.relic.name;
             }